% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blastMenu.R
\name{blast.menu}
\alias{blast.menu}
\title{BLAST sequence and go through menu loops}
\usage{
blast.menu(sequence, database = "pdb", time_out = NULL)
}
\arguments{
\item{sequence}{Sequence as a vector or string to be used to BLAST against}

\item{database}{Menu selected for blast.pdb(). Options: "pdb","nr", or "swissprot". Defaults to "pdb".}

\item{time_out}{Integer indicating time in seconds the blast.pdb() will search before time out occurs. Defaults to}
}
\description{
This function will BLAST the given sequence and give the user an interactive menu to select the best option.
}
