% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blast.menuOptions.R
\name{blast.menuOptions}
\alias{blast.menuOptions}
\title{Make PDB Menu Options}
\usage{
blast.menuOptions(pdb_hit_table, database = "pdb")
}
\arguments{
\item{pdb_hit_table}{Hit table retrieved from bio3d::blast.pdb()}

\item{pdb}{If it searches BLAST PDB. Defaults to true}
}
\description{
This function creates a menu from a pdb_hit_table that is retreived by bio3d's blast.pdb function
}
