% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdb_functions2.R
\docType{data}
\name{bs_output_repeats_pdb2}
\alias{bs_output_repeats_pdb2}
\title{Processed PRC2 RBDmap Data aligned to PDB}
\format{An object of class \code{data.frame} with 22 rows and 6 columns.}
\usage{
data(bs_output_repeats_pdb)
}
\description{
Processed PRC2 RDBmap Data aligned to PDB
}
\keyword{datasets}
