% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdb_functions2.R
\docType{data}
\name{bs_output_repeats_uniprot}
\alias{bs_output_repeats_uniprot}
\title{Processed PRC2 RBDmap Data aligned to Uniprot}
\format{An object of class \code{data.frame} with 22 rows and 10 columns.}
\usage{
data(bs_output_repeats_uniprot)
}
\description{
Processed PRC2 RDBmap Data aligned to Uniprot
}
\keyword{datasets}
