% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdb_functions2.R
\name{check_download_read_pdb}
\alias{check_download_read_pdb}
\title{Check Download Read PDB}
\usage{
check_download_read_pdb(pdb_id)
}
\arguments{
\item{pdb_id}{PDB ID}
}
\description{
This function is a wrapper for the bio3d PDB functions. It checks if the PDB exists before downloading.
}
