% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdb_functions2.R
\name{consolidate_resno_and_resid}
\alias{consolidate_resno_and_resid}
\title{Consolidate resno_and_resid}
\usage{
consolidate_resno_and_resid(resno_and_resid)
}
\arguments{
\item{resno_and_resid}{Output from get_resno_and_resid()}
}
\description{
This function consolidates the get_resno_and_resid() output to get rid of repeating amino acids
}
