% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdb_functions2.R
\name{display_preferred_pdb_structure_menu}
\alias{display_preferred_pdb_structure_menu}
\title{Display Preferred PDB Structure Menu}
\usage{
display_preferred_pdb_structure_menu(
  protein_name,
  fasta_file,
  pwa_score_threshold = 500,
  peptide_sequence = NULL
)
}
\arguments{
\item{protein_name}{Protein name that corresponds to name in fasta_file to extract sequence}

\item{fasta_file}{FASTA file as read by seqinr::read.fasta()}

\item{pwa_score_threshold}{Minimum score to determine a match using Biostrings::pairwiseAlignment()}

\item{peptide_sequence}{An optional peptide sequence to use for recommendations if trying to match a peptide to a structure. Defaults to NULL}
}
\description{
This function creates a menu interface for preferred PDB structure
}
