% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdb_functions.R
\name{do_sr_chain_loop}
\alias{do_sr_chain_loop}
\title{Match query sequence to chain in PDB file}
\usage{
do_sr_chain_loop(protein_name, fasta_seq, pdb_read, pdb_rl,
  pwa_score_threshold = 500, peptide_sequence = NULL, fasta_file = NULL)
}
\arguments{
\item{protein_name}{Protein name that will be used for the console output}

\item{fasta_seq}{Sequence from FASTA file}

\item{pdb_read}{PDB file as loaded by bio3d::seqinr()}

\item{pdb_rl}{Name of PDB file}

\item{pwa_score_threshold}{Minimum score to determine a match using Biostrings::pairwiseAlignment()}

\item{peptide_sequence}{An optional peptide sequence to use for recommendations if trying to match a peptide to a structure. Defaults to NULL}

\item{fasta_file}{Optional fasta file input}
}
\description{
This function will go through all chains in the PDB file to output list of chains that match the sequence
}
