% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdb_functions2.R
\name{fasta.combine}
\alias{fasta.combine}
\title{Combine FASTA files}
\usage{
fasta.combine(fasta_file_names, experiment_directory = NULL,
  fasta_file_directory = NULL,
  fasta_file_in_different_directory = FALSE)
}
\arguments{
\item{fasta_file_names}{List of file names}

\item{experiment_directory}{Experiment directory. If NULL, will use current directory using getwd()}

\item{fasta_file_directory}{If FASTA files are in a different directory. Defaults to NULL.}

\item{fasta_file_in_different_directory}{If FASTA files are in a different directory. Defaults to FALSE.}
}
\description{
This function combines FASTA files from a list of file names
}
