% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdb_functions.R
\name{firstup}
\alias{firstup}
\title{Capitalize first letter}
\usage{
firstup(x)
}
\arguments{
\item{x}{A string}
}
\description{
This function allows you to capitalize the first letter of a string and returns all other characters as lower case.
}
\examples{
string <- 'AAA'
firstup(string)
[1] Aaa
}
\keyword{capitalize}
