% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdb_functions2.R
\name{generate_pdb_lists_from_pdb_csv}
\alias{generate_pdb_lists_from_pdb_csv}
\title{Generate PDB Lists from PDB CSV}
\usage{
generate_pdb_lists_from_pdb_csv(csv_file_name)
}
\arguments{
\item{csv_file_name}{String containing path to csv file}
}
\description{
Generate PDB Lists from PDB CSV
}
