% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdb_functions2.R
\name{get_missing_amino_acids}
\alias{get_missing_amino_acids}
\title{Get missing amino acids/fragments}
\usage{
get_missing_amino_acids(maal, pwa_strings)
}
\arguments{
\item{maal}{Output from renumber_and_identify_missing_aa()}

\item{pwa_strings}{Output from Biostrings::pairwiseAlignment()}
}
\description{
This function finds the missing amino acids between two sequences that have been pairwise aligned
}
