% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pdb_info.R
\name{get_pdb_info}
\alias{get_pdb_info}
\title{Get PDB Info}
\usage{
get_pdb_info(menu_selection, pdb_hit_table)
}
\arguments{
\item{menu_selection}{Numerical selection that has been selected corresponding to the row number of pdb_hit_table}

\item{pdb_hit_table}{Hit table from bio3d::blast.pdb() function}
}
\description{
Get information about PDB selection from menu selection
}
