% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdb_functions.R
\name{get_pwa_ranges}
\alias{get_pwa_ranges}
\title{Get ranges from pairwise alignment results}
\usage{
get_pwa_ranges(pwa_results)
}
\arguments{
\item{pwa_results}{Result from function pairwiseAlignment()}
}
\description{
This function allows you to return ranges (start and end) from pairwise alignment results
}
\examples{
pwa_ranges <- get_pwa_ranges(pwa_results)
}
\keyword{alignment}
\keyword{pairwise}
