% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdb_functions.R
\name{get_pwa_strings}
\alias{get_pwa_strings}
\title{Get string from pairwise alignment results}
\usage{
get_pwa_strings(pwa_results, result_as_vector = FALSE)
}
\arguments{
\item{pwa_results}{Result from function pairwiseAlignment()}

\item{result_as_vector}{If TRUE, returns string as a vector}
}
\description{
This function allows you to return strings from pairwise alignment results
}
\examples{
pwa_strings <- get_pwa_strings(pwa_results)
}
\keyword{alignment}
\keyword{pairwise}
