% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdb_functions.R
\name{get_resno_and_resid}
\alias{get_resno_and_resid}
\title{Get resno and resid from chain in pdb_read}
\usage{
get_resno_and_resid(pdb_read, chain)
}
\arguments{
\item{pdb_read}{PDB file defined by the class in bio3d}

\item{chain}{Character corresponding to chain ID in PDB file}
}
\description{
This function retrieves the residues and corresponding numbers from a specific chain in a PDB file
}
