% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdb_functions2.R
\name{get_uniprot_info_from_proteins_api}
\alias{get_uniprot_info_from_proteins_api}
\title{get uniprot info from proteins api}
\usage{
get_uniprot_info_from_proteins_api(
  uniprot_id,
  type_of_info = c("variation", "feature"),
  select_categories = NULL,
  protein_name = NULL,
  fasta_file = NULL,
  start_and_end_pos = NULL,
  output_xinet_domain_df = FALSE
)
}
\arguments{
\item{uniprot_id}{uniprot_id}

\item{type_of_info}{type_of_info}

\item{select_categories}{select_categories}

\item{protein_name}{protein_name}

\item{fasta_file}{fasta_file}

\item{start_and_end_pos}{start_and_end_pos}

\item{output_xinet_domain_df}{output_xinet_domain_df}
}
\description{
function curreget uniprot info from proteins api
}
