% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdb_functions2.R
\name{get_vector_of_distances_by_pos_freq}
\alias{get_vector_of_distances_by_pos_freq}
\title{Get Vector of Distances by Position Frequency}
\usage{
get_vector_of_distances_by_pos_freq(xl_dataframe)
}
\description{
Get Vector of Distances by Position Frequency
}
