% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdb_functions2.R
\name{get_xyz_coordinates_pdb}
\alias{get_xyz_coordinates_pdb}
\title{Get XYZ coordinates from PDB file}
\usage{
get_xyz_coordinates_pdb(trimmed_pdb)
}
\arguments{
\item{trimmed_pdb}{PDB file loaded by bio3d::read.pdb()}
}
\description{
This function makes a list of the xyz coordinates from a PDB file
}
