% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdb_functions.R
\name{get_xyz_min_max}
\alias{get_xyz_min_max}
\title{Get min and max of xyz coordinates}
\usage{
get_xyz_min_max(xyz_pdb_coords)
}
\arguments{
\item{xyz_pdb_coords}{Output of get_xyz_coordinates_pdb()}
}
\description{
This function gets the min and max of get_xyz_coordinates_pdb() function
}
