% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdb_functions2.R
\name{go_through_menu_loops}
\alias{go_through_menu_loops}
\title{BLAST sequence and go through menu loops}
\usage{
go_through_menu_loops(fasta_sequence_vector, database = "pdb")
}
\arguments{
\item{fasta_sequence_vector}{Sequence as a vector to be used to BLAST against}

\item{pdb}{Menu selected. FALSE uses the 'nr' library. Defaults to TRUE ('pdb' library)}
}
\description{
This function will BLAST the given sequence and give the user an interactive menu to select the best option.
}
