% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbd.xinet.R
\name{is.sequential}
\alias{is.sequential}
\title{Is Sequential}
\usage{
is.sequential(x)
}
\arguments{
\item{x}{List of values to be checked.}
}
\description{
Shows TRUE/FALSE if list contains sequential values
}
