% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdb_functions.R
\name{load_proxl_data}
\alias{load_proxl_data}
\title{Load ProXL Data}
\usage{
load_proxl_data(proxl_data)
}
\arguments{
\item{proxl_data}{File name or data.frame to ProXL data}
}
\description{
This function takes data as downloaded from ProXL's public demo
}
