% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdb_functions.R
\name{make_binding_site_df}
\alias{make_binding_site_df}
\title{Make binding site dataframe}
\usage{
make_binding_site_df(pdb_info, filtered_input_eluate_table_row, fasta_seq)
}
\arguments{
\item{pdb_info}{List of information made from go_through_menu_loops() or do_sr_chain_loop()}

\item{filtered_input_eluate_table_row}{A row of input/eluate table made from make_in make_input_eluate_table()}

\item{fasta_seq}{Vector or string of protein sequence to match the input/eluate information against}
}
\description{
This function makes a row of the RBDMap binding_site_df based on a row of the input/eluate table
}
