% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdb_functions2.R
\name{make_binding_site_rows_for_supp_df}
\alias{make_binding_site_rows_for_supp_df}
\title{make_binding_site_rows_for_supp_df}
\usage{
make_binding_site_rows_for_supp_df(
  proteins_and_intensity_list,
  identifier_is_binding_seq = FALSE
)
}
\arguments{
\item{proteins_and_intensity_list}{proteins_and_intensity_list}

\item{identifier_is_binding_seq}{identifier_is_binding_seq}
}
\description{
make_binding_site_rows_for_supp_df
}
