% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdb_functions.R
\name{make_intensity_output_df}
\alias{make_intensity_output_df}
\title{Make intesnity output dataframe}
\usage{
make_intensity_output_df(sub_shl, prefix)
}
\arguments{
\item{sub_shl}{Subsection of sequence_hit_list, ouput of make_sequence_hit_list()}

\item{prefix}{Prefix that will be used to categorized the subsection in the output}
}
\description{
This function makes a dataframe out of a subsection of the sequence_hit_list
}
