% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_pdb_menu_options.R
\name{make_pdb_menu_options}
\alias{make_pdb_menu_options}
\title{Make PDB Menu Options}
\usage{
make_pdb_menu_options(pdb_hit_table, pdb = TRUE)
}
\arguments{
\item{pdb_hit_table}{Hit table from blast.pdb()}

\item{pdb}{Boolean. If TRUE, uses pdb databases.}
}
\description{
This function makes the PDB menu options
}
