% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdb_functions2.R
\name{make_proteins_and_intensity_list_from_pipeline_output}
\alias{make_proteins_and_intensity_list_from_pipeline_output}
\title{Make Proteins and Intensities List}
\usage{
make_proteins_and_intensity_list_from_pipeline_output(experiment_prefixes,
  protein_to_uniprot_id, fasta_file, fasta_file_directory,
  identifier_is_binding_seq = FALSE)
}
\arguments{
\item{experiment_prefixes}{experiment_prefixes}

\item{protein_to_uniprot_id}{protein_to_uniprot_id}

\item{fasta_file}{fasta_file}

\item{fasta_file_directory}{fasta_file_directory}

\item{identifier_is_binding_seq}{identifier_is_binding_seq}
}
\description{
Internal function used to make the intensities list
}
