% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdb_functions2.R
\name{make_renumbered_pdb_vectors}
\alias{make_renumbered_pdb_vectors}
\title{Make renumbered PDB vectors}
\usage{
make_renumbered_pdb_vectors(resno_and_resid, new_resno_and_resid, maal)
}
\arguments{
\item{resno_and_resid}{Output from get_resno_and_resid() function}

\item{new_resno_and_resid}{Output from consolidate_resno_and_resid() function}

\item{maal}{Output from renumber_and_identify_missing_aa() function}
}
\description{
This function renumbers the PDB file based on a new resno_and_resid
}
