% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdb_functions2.R
\name{make_start_end_pdb_df_output2}
\alias{make_start_end_pdb_df_output2}
\title{Make start and end csv file for make_diff_analysis function}
\usage{
make_start_end_pdb_df_output2(
  generated_pdb_lists,
  output_file_name = "start_and_end"
)
}
\arguments{
\item{generated_pdb_lists}{list() generated from generate_pdb_lists_and_files_from_fasta() or generate_pdb_lists_and_files_pdb_csv()}

\item{output_file_name}{String containing name of output file. Defaults to 'start_and_end'}
}
\description{
This function creates the csv file within the make_diff_analysis function for BS3 XL-MS function
that can be used as input to direct function to correct PDB files
}
