% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdb_functions2.R
\name{make_uv_xlms_visualization}
\alias{make_uv_xlms_visualization}
\title{Make UV XLMS Visualization}
\usage{
make_uv_xlms_visualization(uniprot_fasta_file_numeric,
  output_type = c("xinet", "pymol"))
}
\arguments{
\item{uniprot_fasta_file_numeric}{uniprot_fasta_file_numeric}

\item{output_type}{output_type}
}
\description{
make_uv_xlms_visualization
}
