% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppi.alignPDB.R
\name{ppi.alignPDB}
\alias{ppi.alignPDB}
\title{PPI ALign PDB}
\usage{
ppi.alignPDB(fasta_file, alignIDs = NULL, uniprot2pdb = TRUE)
}
\arguments{
\item{fasta_file}{fasta_file}

\item{alignIDs}{alignIDs}

\item{uniprot2pdb}{uniprot2pdb}
}
\description{
This functions aligns fasta file to PDB
}
