% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdb_functions2.R
\name{ppi.analyze}
\alias{ppi.analyze}
\title{BS3 Differential Analysis Inclusive Function}
\usage{
ppi.analyze(list_of_files, fasta_file, file_type_2d = c("single_dot",
  "linear", "curved"), fasta_names_to_generate_all_2d_structures = NULL,
  csv_pdb_input_file = NULL, category_color_input_file = NULL,
  show_only_real_structures = NULL,
  distance_histogram_name = "distance_histogram.png",
  pymol_file_list_file_name = "pymol_ts_output.pml",
  xlink_df_name = "xlink_dataframe.csv",
  xlink_view_file_name = "xlink_view_df.csv", no_pdb_files = FALSE,
  protein_alternative_names_dict = NULL, pdb_directory = NULL,
  data_input_type = "plink", pdb_numbering = FALSE,
  pdb_match_vector = NULL)
}
\arguments{
\item{list_of_files}{A list of file names containing pLink data to be used for the analysis}

\item{fasta_file}{FASTA file as loaded by seqinr::read.fasta()}

\item{file_type_2d}{The type of file used for PDB creation. "single_dot" creates PDB files with peptides represented by single dots.}

\item{fasta_names_to_generate_all_2d_structures}{List of protein names corresponding to names in FASTA file to automatically make into 2D PDB structures. Defaults to NULL.}

\item{csv_pdb_input_file}{csv_pdb_input_file}

\item{category_color_input_file}{category_color_input_file}

\item{show_only_real_structures}{show_only_real_structures}

\item{distance_histogram_name}{distance_histogram_name}

\item{pymol_file_list_file_name}{pymol_file_list_file_name}

\item{xlink_df_name}{xlink_df_name}

\item{xlink_view_file_name}{xlink_view_file_name}

\item{no_pdb_files}{no_pdb_files}

\item{protein_alternative_names_dict}{protein_alternative_names_dict}

\item{pdb_directory}{pdb_directory}

\item{data_input_type}{data_input_type}
}
\description{
This function combines many of the features of the BS3 XL-MS workflow for ease of use
}
