% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppi.combineData.R
\name{ppi.combineData}
\alias{ppi.combineData}
\title{PPI: Combine Data}
\usage{
ppi.combineData(
  xlink_mega_list,
  fasta_file,
  freq_cutoff = NA,
  score_cutoff = NA,
  cutoff_cond = c("and", "or"),
  protein_alternative_names_dict = NULL,
  category_color_input_file = NULL,
  console_messages = FALSE
)
}
\arguments{
\item{xlink_mega_list}{List made by ppi.loadData()}

\item{fasta_file}{FASTA file as loaded by seqinr::read.fasta()}

\item{freq_cutoff}{Frequency cutoff, will discard values below this number. Defaults to NA.}

\item{score_cutoff}{Score cutoff, will discard values above this number. Defaults to NA.}

\item{cutoff_cond}{Cutoff conditions. If you have values for both freq_cutoff and score_cutoff, this will determine if it will use an "and" or "or" statement.}

\item{protein_alternative_names_dict}{data.frame() or string representing file name. If you have the same proteins represented by different names (NOT RECOMMENDED), you can use this to make sure they will all be combined successfully. Defaults to NULL.}

\item{category_color_input_file}{data.frame() or string representing file name. If you have specific colors you would like to use, use this field.}

\item{console_messages}{Will display console messages about if a match has been made. Primarily for debugging purposes.}
}
\description{
This function combines data and lets you set specifications for how you would like to filter
the data.
}
