% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppi.distAnalysis.R
\name{ppi.distAnalysis}
\alias{ppi.distAnalysis}
\title{PPI Distance Analysis}
\usage{
ppi.distAnalysis(xlink.df.pdb, plot = TRUE)
}
\arguments{
\item{xlink.df.pdb}{Dataframe created by ppi.matchPDB2()}

\item{plot}{If TRUE, will produce the plots. If FALSE, will return a set of dataframes that were used to generate the plots. Defaults to TRUE.}
}
\description{
This function analyzes the distribution of calculated distances of your PDB matches by comparing it to a random distribution of lysines.
}
\author{
Emma Gail
}
