% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdb_functions2.R
\name{ppi.freqCount}
\alias{ppi.freqCount}
\title{Count Frequencies of Sequences in ppi.analyze() output}
\usage{
ppi.freqCount(xlink_df)
}
\arguments{
\item{xlink_df}{Output from ppi.analyze()}
}
\description{
This function counts the frequencies of sequences from ppi.analyze() output in the case of potential repeats of protein positions
}
