% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppi.loadData.R
\name{ppi.loadData}
\alias{ppi.loadData}
\title{PPI: Load Data}
\usage{
ppi.loadData(xldata, fasta_file, file_directory = getwd(), datatype = NULL)
}
\arguments{
\item{xldata}{Name of files or loaded data.frames. If using pLink2 data, it is recommended to use the name of file.}

\item{fasta_file}{FASTA file as loaded by seqinr::read.fasta()}

\item{file_directory}{Directory where XL-MS data is located. Defaults to getwd().}

\item{datatype}{Type of data being loaded. Will automatically detect otherwise. Defaults to NULL.}
}
\description{
This function loads the data from your XL-MS experiments to a format that is readable by ppi.combineData
}
