% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppi.matchPDB.R
\name{ppi.matchPDB}
\alias{ppi.matchPDB}
\title{PPI: Match PDB}
\usage{
ppi.matchPDB(xlink_df, fasta_file, pdb_numbering = FALSE,
  pdb_directory = NULL, pdb_match_vector = NULL,
  csv_pdb_input_file = NULL,
  fasta_names_to_generate_all_2d_structures = NULL,
  file_type_2d = "single_dot")
}
\arguments{
\item{xlink_df}{data.frame() as created by ppi.combineData()}

\item{fasta_file}{FASTA file as loaded by seqinr::read.fasta()}

\item{pdb_numbering}{pdb_numbering}

\item{pdb_directory}{pdb_directory}

\item{pdb_match_vector}{pdb_match_vector}

\item{csv_pdb_input_file}{csv_pdb_input_file}

\item{fasta_names_to_generate_all_2d_structures}{fasta_names_to_generate_all_2d_structures}

\item{file_type_2d}{file_type_2d}
}
\description{
This function matches combined data to PDB files.
}
