% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppi.matchPDB2.R
\name{ppi.matchPDB2}
\alias{ppi.matchPDB2}
\title{PPI Match PDB 2}
\usage{
ppi.matchPDB2(xlink.df, fasta_file, alignIDs, uniprot2pdb = TRUE,
  dist.histogram = TRUE)
}
\arguments{
\item{xlink.df}{xlink.df}

\item{fasta_file}{fasta_file}

\item{alignIDs}{alignIDs}

\item{uniprot2pdb}{uniprot2pdb}

\item{dist.histogram}{dist.histogram}
}
\description{
This function is the newer version
}
