% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppi.matchPDB3.R
\name{ppi.matchPDB3}
\alias{ppi.matchPDB3}
\title{Match PDB 3}
\usage{
ppi.matchPDB3(xlms.df, pdb_numbering)
}
\arguments{
\item{xlms.df}{xlms.df}

\item{pdb_numbering}{pdb_numbering}
}
\description{
Match PDB 3
}
