% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppi.matchUniprot.R
\name{ppi.matchUniprot}
\alias{ppi.matchUniprot}
\title{PPI: Match Protein Sequences to UniProt}
\usage{
ppi.matchUniprot(xlink_df, fasta_file, protein_to_uniprot_id = NULL,
  fasta_directory = NULL, download_fasta = FALSE,
  export_matches = FALSE, remove_na = FALSE, time_out = NULL)
}
\arguments{
\item{xlink_df}{data.frame() as created by ppi.combineData()}

\item{fasta_file}{FASTA file as read by seqinr::read.fasta(). Protein names must correspond to those in xlink_df.}

\item{protein_to_uniprot_id}{Can be either path to table or data.frame. Table containing three columns: ProteinName, UniProtID, and UniProtSeq (optional). When entered, will use these IDs instead of BLAST search.}

\item{fasta_directory}{Directory where the stored FASTA files are located. Defaults to NULL.}

\item{download_fasta}{If TRUE, will download the found UniProt FASTA files as .fasta files to your directory. Defaults to FALSE.}

\item{export_matches}{If TRUE, will export matches from BLAST (swissprot) search as a .csv file for later use. Defaults to FALSE.}

\item{remove_na}{If TRUE, will remove all NAs generated by non-matches in the xlink_df output. Defaults to FALSE.}

\item{time_out}{Integer indicating number of seconds bio3d::blast.pdb() will run before timeout. Defaults to NULL.}
}
\description{
This function matches the protein sequence indeces to those in UniProt if the protein sequences
used were different.
}
