% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppi.pymol.R
\name{ppi.pymol}
\alias{ppi.pymol}
\title{Write PyMOL file for protein-protein interaction data}
\usage{
ppi.pymol(xlink.df, list_of_start_and_end_pdbs = NULL,
  show_only_real_structures = NULL, write_file = FALSE,
  custom.color = FALSE, colors = NULL, color_by = "freq",
  write.df = FALSE, experiment_directory = NULL,
  pdb_numbering = FALSE, pymol_file_list_file_name = "xlink.pml")
}
\arguments{
\item{xlink.df}{data.frame output from ppi.combineData()}

\item{list_of_start_and_end_pdbs}{list_of_start_and_end_pdbs}

\item{show_only_real_structures}{show_only_real_structures}

\item{write_file}{write_file}

\item{custom.color}{custom.color}

\item{colors}{colors}

\item{color_by}{color_by}

\item{write.df}{write.df}
}
\description{
ppi.pymol description
}
