% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdb_functions2.R
\name{ppi.xinet}
\alias{ppi.xinet}
\title{Makes file for xiNET input}
\usage{
ppi.xinet(xlink_df, write_file = TRUE,
  xlink_viewer_csv_file_name = "xlink_viewer.csv", add_color = FALSE)
}
\arguments{
\item{xlink_df}{Dataframe created by make_diff_analysis function}

\item{write_file}{Boolean, TRUE writes file as csv}

\item{xlink_viewer_csv_file_name}{String indicating file name of csv output if write_file = TRUE}
}
\description{
This function writes a file suitable for input into the xiNET xlink viewer website
}
