% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdb_functions2.R
\docType{data}
\name{pymol_color_table}
\alias{pymol_color_table}
\title{PyMol Colors}
\format{An object of class \code{data.frame} with 81 rows and 5 columns.}
\usage{
data(pymol_color_table)
}
\description{
Data on PyMol colors
}
\keyword{datasets}
