% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdb_functions2.R
\name{quick_pwa_from_pdb}
\alias{quick_pwa_from_pdb}
\title{Quick PWA from PDB}
\usage{
quick_pwa_from_pdb(
  query_sequence,
  pdb_read,
  chain,
  use_resid_and_resno = FALSE
)
}
\arguments{
\item{query_sequence}{Sequence to match the PDB sequence against}

\item{pdb_read}{PDB file loaded by bio3d::read.pdb2()}

\item{chain}{Character indicating the chain on the PDB file}

\item{use_resid_and_resno}{Will get the resno_and_resid using get_resno_and_resid() from chain in PDB file rather than the seqres for alignment. Defaults to FALSE.}
}
\description{
This function is a shortcut to do a BioStrings::pairwiseAlignment() with a selected chain in a PDB file
}
