% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdb_functions2.R
\name{quick_resno_and_resid}
\alias{quick_resno_and_resid}
\title{resno_and_resid Shortcut Function}
\usage{
quick_resno_and_resid(pdb_read, chain)
}
\arguments{
\item{pdb_read}{PDB file as loaded by bio3d::read.pdb2()}

\item{chain}{Character indicating the chain on the PDB file}
}
\description{
This function makes the resno_and_resid list directly from a given PDB file and chain
}
