% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbd.alignBS.R
\name{rbd.alignBS}
\alias{rbd.alignBS}
\title{RBD: Align Binding Sequences to Uniprot and PDB}
\usage{
rbd.alignBS(rbd.df, alignIDs, alignTo = c("pdb", "uniprot", "fasta"),
  uniprot2pdb = TRUE, allowPartialBS = FALSE)
}
\arguments{
\item{rbd.df}{rbd.df}

\item{alignIDs}{alignIDs}

\item{alignTo}{alignTo}

\item{uniprot2pdb}{Boolean, defaults to TRUE. This will align to Uniprot sequence before PDB and using the uniprot.PDBmap function to improve alignment accuracy. Highly recommended for PDB alignments.}

\item{allowpartialBS}{Boolean, defaults to FALSE.}
}
\description{
This function aligns the
}
