% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbd.alignBS2.R
\name{rbd.alignBS2}
\alias{rbd.alignBS2}
\title{RBD: Align Binding Sequences to Uniprot and PDB}
\usage{
rbd.alignBS2(
  bs.output,
  alignIDs,
  alignTo = c("pdb", "uniprot", "fasta"),
  uniprot2pdb = TRUE,
  allowPartialBS = FALSE
)
}
\arguments{
\item{bs.output}{Output from rbd.getBSfromIET()}

\item{alignIDs}{alignIDs containing columns of protein names, UniProt and/or PDB identifiers}

\item{alignTo}{What the sequences should be aligned to. Choose between 'pdb' and 'uniprot'}

\item{uniprot2pdb}{Boolean, defaults to TRUE. This will align to Uniprot sequence before PDB and using the uniprot.PDBmap function to improve alignment accuracy. Highly recommended for PDB alignments.}

\item{allowpartialBS}{Boolean, defaults to FALSE. will allow partial matches of binding sequences to PDB/UniProt sequences}
}
\description{
This function aligns the binding sequences
}
\author{
Emma Gail
}
