% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdb_functions2.R
\name{rbd.freqVector}
\alias{rbd.freqVector}
\title{Get Frequency Vector for RBDmap}
\usage{
rbd.freqVector(bs_output, name_by = "pro_name", heatmap = TRUE,
  db_selection = NULL, colors = c("#d0d0d0", "#2f5ac6", "#e50000"))
}
\arguments{
\item{bs_output}{bs_output}

\item{name_by}{name_by}

\item{heatmap}{heatmap}

\item{db_selection}{db_selection}

\item{colors}{colors}
}
\description{
This function gets the frequency vector
}
