% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdb_functions2.R, R/rbd.getBindingSeq.R
\name{rbd.getBindingSeq}
\alias{rbd.getBindingSeq}
\title{Get Binding Sequence from RBDmap Data}
\usage{
rbd.getBindingSeq(ms_sequence, protease, sequence_for_alignment,
  protein_name = NULL, database_name = NULL, database_id = NULL,
  cleave_offset = 0, last_aa = NULL, include_ambiguous = FALSE,
  proteolytic_fragments = FALSE)

rbd.getBindingSeq(ms_sequence, protease, sequence_for_alignment,
  protein_name = NULL, database_name = NULL, database_id = NULL,
  cleave_offset = 0, last_aa = NULL, include_ambiguous = FALSE,
  proteolytic_fragments = FALSE)
}
\arguments{
\item{ms_sequence}{The input sequence from MS, should be loaded as a string}

\item{protease}{The protease used for the experiment, either 'ArgC' or 'LysC'}

\item{sequence_for_alignment}{The sequence to find the binding sequence in. The ms_sequence must be within the sequence_for_alignment or NULL will be returned.}

\item{protein_name}{Name of the protein to which the sequence_for_alignment belongs. Defaults to NA}

\item{database_name}{Name of the database used: 'PDB','UniProt' or 'FASTA' (for the input FASTA sequence). Defaults to NA}

\item{database_id}{Identifier for the database. Defaults to NA}

\item{cleave_offset}{Cleave offset for the sequence. Defaults to 0}

\item{ms_sequence}{The input sequence from MS, should be loaded as a string}

\item{protease}{The protease used for the experiment, either 'ArgC' or 'LysC'}

\item{sequence_for_alignment}{The sequence to find the binding sequence in. The ms_sequence must be within the sequence_for_alignment or NULL will be returned.}

\item{protein_name}{Name of the protein to which the sequence_for_alignment belongs. Defaults to NULL}

\item{database_name}{Name of the database used: 'PDB','UniProt' or 'FASTA' (for the input FASTA sequence). Defaults to NULL}

\item{database_id}{Identifier for the database. Defaults to NULL}

\item{cleave_offset}{Cleave offset for the sequence. Defaults to 0}
}
\description{
This funtion gets the binding sequence from RBDmap data

This funtion gets the binding sequence from RBDmap data
}
