% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbd.getBindingSeq3.R
\name{rbd.getBindingSeq3}
\alias{rbd.getBindingSeq3}
\title{RBD: Get Binding Sequence 3}
\usage{
rbd.getBindingSeq3(trypticPeptide, enzyme, sourceSequence,
  cleave_offset = 4)
}
\arguments{
\item{trypticPeptide}{trypticPeptide}

\item{enzyme}{enzyme}

\item{sourceSequence}{sourceSequence}

\item{cleave_offset}{cleave_offset}
}
\description{
This function gets the binding sequence
}
