% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdb_functions.R
\name{rbd.getBSfromIET}
\alias{rbd.getBSfromIET}
\title{Get Binding Sequences from Input/Eluate Table}
\usage{
rbd.getBSfromIET(filtered_input_eluate_table, fasta_file, align_to = "PDB",
  protein_to_uniprot_id = NULL, protein_dict = NULL,
  include_ambiguous = FALSE)
}
\arguments{
\item{filtered_input_eluate_table}{filtered_input_eluate_table}

\item{fasta_file}{fasta_file}

\item{align_to}{align_to}

\item{protein_to_uniprot_id}{protein_to_uniprot_id}

\item{protein_dict}{protein_dict}

\item{include_ambiguous}{include_ambiguous}
}
\description{
This function gets the binding sequences from the input eluate table
}
