% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdb_functions2.R
\name{rbd.makeIETable}
\alias{rbd.makeIETable}
\title{Make input and eluate table}
\usage{
rbd.makeIETable(sequence_hit_list, prefixes, proteases = c("ArgC",
  "LysC"), output_file_name = "pipeline_output_file.csv",
  experiment_directory = NULL)
}
\arguments{
\item{sequence_hit_list}{Output from rbd.makeSeqHitList() function}

\item{prefixes}{Prefix in file name identifying the experiments}

\item{proteases}{Proteases used in the experiment. Defaults to c('ArgC','LysC')}

\item{outline_file_name}{Name of output file. Defaults to 'pipeline_output_file.csv'}
}
\description{
This function makes an input and eluate table based on the make_sequence_hit_list() function
}
\author{
Emma Gail
}
