% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdb_functions.R
\name{rbd.menuDBSearch}
\alias{rbd.menuDBSearch}
\title{Menu DB Search}
\usage{
rbd.menuDBSearch(input_sequence, fasta_file, protein_name, pdb_info = NULL)
}
\arguments{
\item{input_sequence}{input_sequence}

\item{fasta_file}{fasta_file}

\item{protein_name}{protein_name}

\item{pdb_info}{pdb_info}
}
\description{
Menu DB Search
}
