% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbd.xinet.R
\name{rbd.xinet}
\alias{rbd.xinet}
\title{RBDmap: xiNET Output}
\usage{
rbd.xinet(bs.output, colors = c("#d0d0d0", "#2f5ac6", "#e50000"),
  write.file = TRUE, file.name = "rbd_xinet.csv")
}
\arguments{
\item{bs.output}{Output from function rbd.getBSfromIET()}

\item{colors}{Colors to be used for the annotation file. Can be an RColorBrewer palette name, virdis palette name, or list of hexcodes.}

\item{write.file}{If TRUE, will write to a .csv file}

\item{file.name}{File name for the written out file.}
}
\description{
This function outputs an annotation file for import into the web server xiNET. Designed to be used in conjunction with output from ppi.xinet().
}
