% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdb_functions.R
\name{renumber_and_identify_missing_aa}
\alias{renumber_and_identify_missing_aa}
\title{Renumber and identify missing amino acids in pairwise alignment results}
\usage{
renumber_and_identify_missing_aa(pwa_strings, start_difference)
}
\arguments{
\item{start_difference}{Difference between starting amino acid in subject and pattern alignments}

\item{pwa_results}{Result from function pairwiseAlignment()}
}
\description{
This function allows you to renumber and identify missing amino acids in pairwise alignment results
}
\examples{
renumbered_missing_amino_acids <- renumber_and_identify_missing_aa(pwa_strings,start_difference)
}
\keyword{alignment}
\keyword{pairwise}
