% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdb_functions2.R
\name{renumber_binding_site_df_from_uniprot_fasta}
\alias{renumber_binding_site_df_from_uniprot_fasta}
\title{Renumber binding site df from uniprot fasta}
\usage{
renumber_binding_site_df_from_uniprot_fasta(binding_site_df,
  protein_to_uniprot_id, fasta_file, fasta_file_directory = NULL,
  use_pipeline_output_df = FALSE)
}
\arguments{
\item{binding_site_df}{binding_site_df}

\item{protein_to_uniprot_id}{protein_to_uniprot_id}

\item{fasta_file}{fasta_file}

\item{fasta_file_directory}{fasta_file_directory}

\item{use_pipeline_output_df}{use_pipeline_output_df}
}
\description{
Renumber binding site df from uniprot fasta
}
