% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdb_functions.R
\name{split_sequences2}
\alias{split_sequences2}
\title{Split sequences from plink output}
\usage{
split_sequences2(xlink_string, proteins = TRUE)
}
\arguments{
\item{xlink_string}{String of BS3 XL data made by make_plink2_master_list()}

\item{proteins}{TRUE/FALSE if string needs to be split by ":"}
}
\description{
This function splits sequences from the plink output function
}
