% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniprotfasta.R
\name{uniprot.fasta}
\alias{uniprot.fasta}
\title{Fetch FASTA sequence from Uniprot}
\usage{
uniprot.fasta(uniprot_id, download_fasta = FALSE,
  console_message = TRUE, fasta_directory = getwd())
}
\arguments{
\item{uniprot_id}{uniprot_id}

\item{download_fasta}{download_fasta}

\item{console_message}{console_message}

\item{fasta_directory}{fasta_directory}
}
\description{
This function fetches the FASTA sequence from UniProt
}
