% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniprot.PDBmap.R
\name{uniprot.PDBmap}
\alias{uniprot.PDBmap}
\title{UniProt PDBmap}
\usage{
uniprot.PDBmap(pdb_id, resno, chain = NULL, uniprot_id = NULL,
  output = c("pdb", "uniprot"))
}
\arguments{
\item{pdb_id}{Valid ID of a PDB accession number within RCSB}

\item{resno}{Residue number(s) to match against the PDB/UniProt number.}

\item{chain}{Chain ID within the PDB to match. Needed if matching PDB residue number to UniProt ID. Defaults to NULL}

\item{uniprot_id}{UniProt ID of protein. Needed if matching UniProt sequence to residue within PDB}

\item{output}{Choose between 'pdb' or 'uniprot' for which ID the output residue number should correspond to}
}
\description{
This function is able to connect with the RCSB API to be able to match a residue number of either a PDB chain or UniProt canonical sequence to its match.
}
\author{
Emma Gail
}
