% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdb_functions2.R
\docType{data}
\name{xlink.df.filt.pdb}
\alias{xlink.df.filt.pdb}
\title{Processed PRC2 PPI Data aligned to PDB}
\format{An object of class \code{data.frame} with 29 rows and 17 columns.}
\usage{
data(xlink.df.filt.pdb)
}
\description{
Processed PRC2 PPI Data aligned to PDB
}
\keyword{datasets}
