% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdb_functions2.R
\docType{data}
\name{xlink.df.uniprot}
\alias{xlink.df.uniprot}
\title{Processed PRC2 PPI Data aligned to Uniprot}
\format{An object of class \code{data.frame} with 70 rows and 14 columns.}
\usage{
data(xlink.df.uniprot)
}
\description{
Processed PRC2 PPI Data aligned to Uniprot
}
\keyword{datasets}
