% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppi.matchPDB2.R
\name{ppi.matchPDB2}
\alias{ppi.matchPDB2}
\title{PPI Match PDB 2}
\usage{
ppi.matchPDB2(xlink.df, fasta_file, alignIDs, uniprot2pdb = TRUE,
  dist.histogram = TRUE)
}
\arguments{
\item{xlink.df}{A data.frame made by ppi.combineData()}

\item{fasta_file}{Name of fasta file or loaded fasta file by seqinr::read.fasta().}

\item{alignIDs}{A data.frame containing the columns "ProteinName", "UniProtID", and "PDB"}

\item{uniprot2pdb}{If TRUE, will align to the UniProt sequence before aligning to the PDB. This parameter should be selected if the sequences used are not exactly UniProt (such as a slightly different N-terminal), but are relatively similar.}

\item{dist.histogram}{dist.histogram}
}
\description{
This function is the newer version
}
