% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbd.freqVector.R
\name{rbd.freqVector}
\alias{rbd.freqVector}
\title{Get Frequency Vector for RBDmap Data}
\usage{
rbd.freqVector(bs_output, name_by = "pro_name", heatmap = TRUE,
  db_selection = NULL, colors = c("#d0d0d0", "#2f5ac6", "#e50000"),
  save_plot = TRUE)
}
\arguments{
\item{bs_output}{data.frame() output from rbd.getBSfromIET() or rbd.getBSfromDF()}

\item{name_by}{Name of column in bs_output that will correspond to where the binding sites will be measured}

\item{heatmap}{If TRUE, a heatmap will be produced. If FALSE, the data.frame used for the heatmap will be returned}

\item{db_selection}{Database to be used for the heatmap (e.g. FASTA, PDB, or UniProt)}

\item{colors}{Colors to be used for the heatmap, if TRUE. Can be a list of hexcodes/standard R colors or RColorBrewer palette}

\item{save_plot}{If TRUE, will save the plot to your working directory}
}
\description{
This function gets the frequency vector of the binding sequences detected
}
\author{
Emma Gail
}
