% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbd.getBSfromDF.R
\name{rbd.getBSfromDF}
\alias{rbd.getBSfromDF}
\title{RBD: Get binding sequences from Dataframe}
\usage{
rbd.getBSfromDF(rbd.df, fasta = NULL)
}
\arguments{
\item{rbd.df}{rbd.df}

\item{fasta}{Name of fasta file or loaded fasta file by seqinr::read.fasta(). Defaults to NULL.}
}
\description{
Function description
}
