% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbd.getBSfromIET.R
\name{rbd.getBSfromIET}
\alias{rbd.getBSfromIET}
\title{Get Binding Sequences from Input/Eluate Table}
\usage{
rbd.getBSfromIET(
  filtered_input_eluate_table,
  fasta_file,
  align_to = "PDB",
  protein_to_uniprot_id = NULL,
  protein_dict = NULL,
  include_ambiguous = FALSE
)
}
\arguments{
\item{filtered_input_eluate_table}{Output from rbd.makeIETable()}

\item{fasta_file}{FASTA file from seqinr::read.fasta()}

\item{align_to}{What source to align the BS to: "PDB","FASTA","UniProt"}

\item{protein_to_uniprot_id}{Table containing protein_id,uniprot_id,pdb_id}

\item{protein_dict}{protein_dict}

\item{include_ambiguous}{include_ambiguous}
}
\description{
This function gets the binding sequences from the input eluate table
}
