% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdb_functions2.R
\name{rbd.makeSeqHitList}
\alias{rbd.makeSeqHitList}
\title{Make sequence hit list}
\usage{
rbd.makeSeqHitList(fasta_file, experiment_directory = NULL,
  file_format = "csv", cutoff_score = 20, RK.cleavage = TRUE)
}
\arguments{
\item{fasta_file}{FASTA file as loaded by seqinr::read.fasta()}

\item{experiment_directory}{Directory where experiment files are. If NULL, defaults to current directory using getwd().}

\item{file_format}{Can be either 'csv' or 'xlsx', the suffix to the experiment files}

\item{cutoff_score}{Cutoff for Andromeda score. Defaults to 20.}
}
\description{
This function creates a hit list of significant XL in UV XL-MS dataset
}
\author{
Emma Gail
}
