% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniprotfasta.R
\name{uniprot.fasta}
\alias{uniprot.fasta}
\title{Fetch FASTA sequence from Uniprot}
\usage{
uniprot.fasta(uniprot.id, download.fasta = FALSE,
  console.message = TRUE, fasta.directory = getwd())
}
\arguments{
\item{uniprot.id}{Uniprot ID of sequence wanted in the form "ID-isoform". Leave out isoform if using the canonical sequence.}

\item{download.fasta}{If TRUE, will download the FASTA file to your chosen directory. Defaults to FALSE.}

\item{console.message}{If TRUE, will display the name of the FASTA file when being loaded into R in the console. Defaults to TRUE.}

\item{fasta.directory}{Directory chosen for your downloaded FASTA file. Defaults to current directory.}
}
\description{
This function fetches the FASTA sequence directly from the UniProt .fasta page
}
\author{
Emma Gail
}
